Revert "Remove "the first steps" and "the perfect setup" sections"
This reverts commit 0a2a03079b15442bb4b1f30fbbd3d6698a848ecf.

I removed the perfect setup link because the contents of that page
were "lost" during the wiki transition, but now I noticed that Alan
Coopersmith had restored that page, so I'm reverting my change.

I removed the first steps link, because I don't really like the
contents of that page, but now I'm reverting that change too, because
I don't want to start the discussion whether that is a good idea or
not (I was prepared to have that discussion as part of the discussion
whether removing the perfect setup link is a good idea, but now that
the perfect setup has been restored, the situation is different).