Add filesystem option, and rework paths
We need to be able to support reading files from partitions other than $BOOT
and so need a mechanism for providing a pointer to the correct partition.
Add the filesystem keyword and define its behaviour on BIOS and UEFI systems,
and change the text for keywords that expect a filename to match.