Substantially rewrite a lot of pages
Rework our front page to be a lot more useful, as well as links to
getting involved and starting new projects.

Rewrite our mission statement to be more consistent with what we've
actually interpreted it as for the last ten years. This probably
requires further tweaking.

Make a first cut at improving the specs page, though this needs a lot
more work.

Mostly rewrite infrastructure-related pages to semi-accurately reflect
the current state of things.

LightDM has fully moved to GitHub now, so have their page just be a
redirect rather than an orphan.